Stabilize tests
By moving the reset in the teardown with some breathing we make sure
mock server has time to finish processing the requests from test
because there seem to be some race condition when verifying calls.

Since the setup doesn't do the reset anymore we need to stall until
mock server is ready to serve so the waiting was pushed in the test
runner checking with curl after the compose up is done.